Move SearchViewModel to KMM
