--- androidApp/src/main/java/dev/johnoreilly/confetti/search/SearchModule.kt
@@ -1,8 +1,0 @@
-package dev.johnoreilly.confetti.search
-
-import org.koin.androidx.viewmodel.dsl.viewModelOf
-import org.koin.dsl.module
-
-fun searchModule() = module {
-    viewModelOf(::SearchViewModel)
-}
